// src/types/organization/organization.d.ts
export interface organizationType {
	id?: number;
	name: string;
	registration_number: string;
	industry_type: string;
	sector: string; // ngo, private, federal
	number_of_employees: string;
	company_website: string;
	plan_coverage_type: string;
	preferred_start_date: string;
	preferred_end_date: string;
	country_of_origin: string;
	country: string;
	street_address: string;
	city: string;
	region: string;
	kifle_ketema: string;
	phone_number: string;
	email_address: string;
	contact_person_first_name: string;
	contact_person_middle_name: string;
	contact_person_last_name: string;
	contact_person_position: string;
	contact_person_phone_number: string;
	contact_person_email_address: string;
}